Support for Hardware Version 2.2 (#27)
* Remapped all pins for v2 boards

* Moving DMA Stream for SPI3 to prevent overlap with I2C2 and correcting LED colors

* Moving old imu driver

* Adjusting constants/config for new boards

* Driver Updates (Not working fully yet but avoids address collision)

* GDB Multiarch replaced specific gdb versions

* Removing voltage readings

* Tools update for v2 boards

* Fixed accelerometer and temperature sensor drivers


* Fixed ADC parameters for new board revision


* WIP: Fixed ADC structure for new ADC channels


* More consistency in ADC channels


* Recorder updates

* Fixed inverted ADC readings


* Changed to center aligned PWM.

While the frequency constant was doubled because of the center aligned feature the PWM will still be at 20kHz frequency


* Calibration script fixes (add sleep and increase timeout)


* Updated recorder element count

* Changed ADC sampling to be relative to falling edge instead of PWM center


* Revert "Changed ADC sampling to be relative to falling edge instead of PWM center"

This reverts commit a06cbe19c0f1e92e3d5ec617d9e5343458775861.

* Removed PWM limit, halved update rate from timer

* More conservative delay for bootloader exit


* Fixing more tools

* Fixed current readings for low PWM duty cycles

Clamped PWM duty cycle between 0.05 and 0.95, added simultaneous
sampling of all the current channels


* Getting gravity to match output

* imu test now produces correct gravity scaled vector

+++ firmware/include/state.h
@@ -40,9 +40,9 @@ struct Results {
   float ib = 0;                             // current into phase B (amperes)
   float ic = 0;                             // current into phase C (amperes)
 
-  int32_t xl_x = 0;                         // X-acceleration in milli-g's
-  int32_t xl_y = 0;                         // Y-acceleration in milli-g's
-  int32_t xl_z = 0;                         // Z-acceleration in milli-g's
+  int16_t xl_x = 0;                         // X-acceleration in milli-g's
+  int16_t xl_y = 0;                         // Y-acceleration in milli-g's
+  int16_t xl_z = 0;                         // Z-acceleration in milli-g's
 
   float temperature = 0;                    // Temperature in degrees Celsius
